Other compiler warning fixes: this is another chunk of compiler warning fixes in LTP tests, related to printf format strings. They have multiple causes: - most of them are caused by %d used for the TEST_RETURN variable (as in a previous patch), - off_t are cast to (int64_t) and PRId64 output specifier is used for them, - other types are cast to (intmax_t) or (uintmax_t) and use the %jd or %ju (C99) output specifier. I've used this solution because it's generally correct, easy to use solution that is portable and (IMHO) more readable than using PRId64/PRIu64 and a cast (which you generally can't avoid). This way, there's only a cast, - in ioperm test, address should be declared as unsigned long (according to ioperm(2)), - other little fixes (foo vs. *foo, %S vs. %s, %0 vs. %o). Signed-off-by: Jiri Palecek <jpalecek@web.de>.

+++ testcases/kernel/mem/hugetlb/hugeshmctl/hugeshmctl01.c
@@ -352,8 +352,8 @@ func_stat()
 	 * that memory so the attaches equal N_ATTACH + stat_time (1).
 	 */
 	if (!fail && buf.shm_nattch != N_ATTACH + stat_time) {
-		tst_resm(TFAIL, "# of attaches is incorrect - %d",
-			 buf.shm_nattch);
+		tst_resm(TFAIL, "# of attaches is incorrect - %lu",
+			 (unsigned long)buf.shm_nattch);
 		fail = 1;
 	}
